Correct sql creation to remove extranous columns
The table creation specified columns 'src', 'dst', 'channel', and
'dstchannel' that should not exist.  With those channels in the
table, the Asterisk 13.7 code code in cel/cel_odbc.c will fail
because it only inserts the following columns:

	eventtype,eventtime,cid_name,cid_num,cid_ani,cid_rdnis,
	cid_dnid,exten,context,channame,appname,appdata,amaflags,
	accountcode,uniqueid,linkedid,peer,userdeftype,extra

Removing these columns brings the table in line with the Asterisk
documentation as well.